feat: migrate to justfile

+++ install-fonts.sh
@@ -1,5 +1,19 @@
-# JetBrainsMono nerd font
-cd /Library/Fonts
-wget https://github.com/ryanoasis/nerd-fonts/releases/download/v3.1.1/JetBrainsMono.zip
-unzip JetBrainsMono.zip
-rm JetBrainsMono.zip
+FONTS_DIR="/Library/Fonts"
+
+function isFontInstalled() {
+    if [ $(ls -a $FONTS_DIR | grep -i "$1" | wc -l | tr -d ' ') -gt 0 ]; then
+        echo true
+    else
+        echo false
+    fi
+}
+
+if [ ! $(isFontInstalled "JetBrains") ]; then
+    cd $FONTS_DIR
+    echo "Installing JetBrains Mono nerd font..."
+    wget https://github.com/ryanoasis/nerd-fonts/releases/download/v3.1.1/JetBrainsMono.zip
+    unzip JetBrainsMono.zip
+    rm JetBrainsMono.zip
+else
+    echo "JetBrains Mono already installed, skipping..."
+fi
